feat(triple-web): 폴더 구조 개선
- contexts 노출하지 않음
- types 분리
- useContext 훅을 기본으로 같은 파일에 정의하고 노출할 훅만 hooks에서 re-export
