+++ packages/triple-web/src/contexts/client-app.ts
@@ -1,19 +1,17 @@
-import { createContext } from 'react'
+import { createContext, useContext } from 'react'
 
-export const enum ClientAppName {
-  iOS = 'Triple-iOS',
-  Android = 'Triple-Android',
+import { ClientAppValue } from '../types'
+
+export const ClientAppContext = createContext<ClientAppValue | undefined>(
+  undefined,
+)
+
+export function useClientApp() {
+  const context = useContext(ClientAppContext)
+
+  if (context === undefined) {
+    throw new Error('ClientAppContext가 없습니다.')
+  }
+
+  return context
 }
-
-export type ClientApp = {
-  metadata: {
-    name: ClientAppName
-    version: string
-  }
-  device: {
-    autoplay: 'always' | 'wifi_only' | 'never'
-    networkType: 'wifi' | 'cellular' | 'unknown'
-  }
-} | null
-
-export const ClientAppContext = createContext<ClientApp | undefined>(undefined)
