+++ packages/triple-web/src/hooks/client-app/use-client-app.ts
@@ -1,13 +1,1 @@
-import { useContext } from 'react'
-
-import { ClientAppContext } from '../../contexts/client-app'
-
-export function useClientApp() {
-  const context = useContext(ClientAppContext)
-
-  if (context === undefined) {
-    throw new Error('ClientAppContext가 없습니다.')
-  }
-
-  return context
-}
+export { useClientApp } from '../../contexts'
