+++ packages/triple-web/src/hooks/session/use-session.ts
@@ -1,13 +1,1 @@
-import { useContext } from 'react'
-
-import { SessionStateContext } from '../../contexts'
-
-export function useSession() {
-  const context = useContext(SessionStateContext)
-
-  if (context === undefined) {
-    throw new Error('')
-  }
-
-  return context
-}
+export { useSession } from '../../contexts'
