+++ packages/triple-web/src/index.ts
@@ -1,4 +1,4 @@
-export * from './contexts'
+export * from './components'
 export * from './hooks'
-export * from './utils'
+export * from './types'
 export * from './triple-web'
