+++ packages/triple-web/src/types/client-app.ts
@@ -1,0 +1,15 @@
+export const enum ClientAppName {
+  iOS = 'Triple-iOS',
+  Android = 'Triple-Android',
+}
+
+export type ClientAppValue = {
+  metadata: {
+    name: ClientAppName
+    version: string
+  }
+  device: {
+    autoplay: 'always' | 'wifi_only' | 'never'
+    networkType: 'wifi' | 'cellular' | 'unknown'
+  }
+} | null
