+++ packages/triple-web/src/types/event-metadata.ts
@@ -1,0 +1,3 @@
+export interface EventMetadataValue {
+  [key: string]: string
+}
