+++ packages/triple-web/src/types/event-tracking.ts
@@ -1,0 +1,7 @@
+export interface EventTrackingValue {
+  page: {
+    label: string
+    path: string
+  }
+  onError?: (error: Error) => void
+}
