+++ packages/triple-web/src/types/modal.ts
@@ -1,0 +1,19 @@
+import { MutableRefObject } from 'react'
+
+import { EventTrackingValue } from './event-tracking'
+
+export interface LoginCtaModalRef {
+  returnUrl?: string
+  triggeredEventLabel?: string
+}
+
+export interface TransitionModalRef {
+  deepLink?: string
+  onActionClick?: () => void
+}
+
+export interface ModalValue {
+  loginCtaModalRef: MutableRefObject<LoginCtaModalRef>
+  transitionModalRef: MutableRefObject<TransitionModalRef>
+  eventTrackingContextForkRef: MutableRefObject<EventTrackingValue | undefined>
+}
