+++ packages/triple-web/src/types/user-agent.ts
@@ -1,0 +1,3 @@
+import { IResult } from 'ua-parser-js'
+
+export type UserAgentValue = IResult
